update sweepRoot usage

+++ condorMergingTools/makeListsForMerging.py
@@ -100,6 +100,11 @@ def getGoodRootFiles(datapath,outpath):
                 
         #checks the files with sweeproot before copying. If it finds a bad file, It skips the file and does not add the events to the total # of events
         print 'Checking File /hadoop' + dcachePrefix + path + ' for integrity'
+        if os.path.isfile('sweepRoot') == False:
+            print "sweepRoot executable does not exist! Exiting."
+            print "sweepRoot source exists in sweepRoot directory."
+            sys.exit()
+
         cmd = "./sweepRoot -o Events /hadoop/" + dcachePrefix + path #outpath + '/temp/' + fname + ' 2> /dev/null'
         output = commands.getoutput(cmd).split('\n')
 
@@ -242,14 +247,20 @@ if datapath.find("pnfs") != -1 or datapath.find("hadoop") != -1:
         sys.exit()
 
         
+##now make the sweeproot macro if it doesn't exist
+##sweepRoot should already exist when being called from AutoTuple
+if os.path.isfile('sweepRoot') == False:
+    cmd = 'cp ../sweepRoot/Makefile .'
+    commands.getstatusoutput(cmd)
+    cmd = 'cp ../sweepRoot/sweepRoot.C .'
+    commands.getstatusoutput(cmd)
+    cmd = 'make '
+    commands.getstatusoutput(cmd)
 
-##now make the sweeproot macro
-cmd = 'cp ../sweepRoot/Makefile .'
-commands.getstatusoutput(cmd)
-cmd = 'cp ../sweepRoot/sweepRoot.C .'
-commands.getstatusoutput(cmd)
-cmd = 'make '
-commands.getstatusoutput(cmd)
+if os.path.isfile('sweepRoot') == False:
+    print "sweepRoot executable does not exist! Exiting."
+    print "sweepRoot source exists in sweepRoot directory."
+    sys.exit()
 
 
 goodRootFiles = []
